Modernize update alert (#2737)
This changes the UI for the update alert to have the app icon be more stacked / inlined with the UI, refines the box view for release notes, and removes some title labels that aren't strictly needed. For Tahoe, we opt into using large size buttons which get a capsule shape.

Also this changes adds a new app icon, which is squircle shaped, for the Sparkle test when running on Tahoe.