Add hardware requirements element (#2797)
The first requirement we support is arm64 which allows updates to mark themselves as Apple silicon only.

The message that shows up when users check for updates tells them their Mac is too old and a newer Apple silicon Mac is required for new updates. This introduces a new localizable string.

This requirement is also inferred if the minimum system version is >= 27.0

generate_appcast is also updated to handle a new branch for hardwareRequirements element, and for automatically populating it if needed.

Fixes #2796