Make Debug builds use same time interval settings as Release (#2805)
The Sparkle Test App will opt into a private setting that changes some of the default and minimum values for Debug builds.

This also removes the "WARNING: This is running a Debug build of Sparkle 2; don't use this in production!" log in Debug builds, and removes some unused constants internal to Sparkle.