Add allowsAutomaticUpdates property (#2809)
This allows us to determine if automatic downloading/installing of updates option (typically a checkbox) should be enabled or not. This property is based on whether or not automatically checking of updates is enabled, or the SUAllowsAutomaticUpdates option (if specified by developer for custom needs).

The update alert now updates hiding/unhiding the automatic downloading checkbox if the automatic update checking option changes.

The test application is now updated to use -[SPUUpdater allowsAutomaticUpdates] for enabling/disabling the automatic downloads update checkbox.